Feat: Updated scoring function
Updated the scoring function to rescore the query list if a prompt is
entered and subsequently cleared. i.e. Reset it back to an ordered list
by the query timestamp.

It does recompute the list each time so it's probably not ideal
performance wise, should look at computing the known chronological list
once and cache it ( either each time the extension is run or some other
interval ).

It now also only looks for full matches when a prompt is entered, I
don't really see the benefit in looking for partial matches currently.