Feat: Update default select behaviour
When a query is selected, a new scratch buffer is opened and switched
to, it contains the query, with the filetype set to sql for
highlighting.

Perhaps there will be a desire to either:

Open the query in a buffer
Copy the query to the register
Both!

This should be possible to do via general query behaviour which can be
set in config at the top level for all query pickers.